Mass update to the project page.
Noticed during HCDE 547 while talking websites that this part of my website was very out of date!
To remove some copy errors and redundancy, I made the projects page into more of a list and each project has more of it's own info.
Also noticed that the embedded docs on each project page looked stupid so I just linked to pages instead.
Lastly, renames a bunch of the files and made separate folders for each project.